fix: address code review findings — stale locations, event consistency, overflow safety
- PaginatedMenu.open() now fires MenuOpenEvent for consistency
- HologramGUI/NpcGUI teleport lambdas fetch fresh location at click time
- MenuTickTask tickCount changed from int to long to prevent overflow
- Menu.refresh() accepts long tickCount
- Null guard on hologram entity UUID in show/hide convenience methods

+++ src/main/java/net/axther/serverCore/gui/Menu.java
@@ -125,7 +125,7 @@ public class Menu {
      * @param player    the player whose inventory to refresh
      * @param tickCount global tick counter used for cycle item rotation
      */
-    public void refresh(Player player, int tickCount) {
+    public void refresh(Player player, long tickCount) {
         Inventory topInventory = player.getOpenInventory().getTopInventory();
         if (topInventory == null) return;
 
@@ -138,7 +138,7 @@ public class Menu {
             List<ItemStack> cycleItems = item.getCycleItems();
             if (cycleItems != null && !cycleItems.isEmpty()) {
                 int interval = Math.max(1, item.getCycleInterval());
-                int index = (tickCount / interval) % cycleItems.size();
+                int index = (int) ((tickCount / interval) % cycleItems.size());
                 topInventory.setItem(slot, cycleItems.get(index));
                 continue;
             }
